Implemented squad tab
Implemented the squad screen using Koin for dependency injection, MVVM architecture with LiveData and Repository pattern.
Added all the resources and created the classes needed to implement the tabLayout and the squad tab according to the design.
Some of the images are added as .png file instead of .svg because after exported them from Figma and imported into Android Studio there were some errors like `ERROR @ line 7: <pattern> is not supported ERROR @ line 13: <image> is not supported`
Created two diferent themes - one for UCL version and one for UEL version.

*Are there any specific API/Data concerns to mention?
Setup the retrofit in order to retrieve the date from the API, but for the demo part, added in the assets directory a .json file with all the data regarding how the models should look.
The assumption is that the images related to squad will be also received from the backend and the URL needs to be added on the backed model, but for this excercise the images are loaded from locally.
Regarding `*Player list B` the assumption is that there should be another list with all the bench players. Clicking on this footer will expand this list.

*Please explain the UI/API contract to follow in case another developer needs to fulfill the UI with real data.
The viewModel retrieves the data from the repository. The repository should retrieve the data from the API or in this case will get the mocked data from the .json file.
The viewModel maps all the backend models to others UI models which are used by data binding.
The view - activity observes for models list from viewModel and updates the UI based on the model changes.

*Would you split this ticket into more granular pieces? How would you do it if so?
Yes, the idea was to split it into two tasks - one task for the squad tab(the list with all the data) and one task for the header with all the tabs.

*What’s the unit test coverage followed on this PR?
Since the viewModel contains only the lists with the data retrieved from the API/demo and the mappers methods, all the unit tests will be made whem more busines logic will be required.